+++ backend/services/scaling-service/scaling_algorithms/helpers/capacity.go
@@ -14,7 +14,7 @@ import (
 
 // ComputeRealMandelboxCapacity is a helper function to compute the real mandelbox capacity.
 // Real capcity is the number of free mandelboxes available on instances with active status.
-func ComputeRealMandelboxCapacity(imageID string, activeInstances []subscriptions.Instance) int {
+func ComputeRealMandelboxCapacity(imageID string, activeInstances []subscriptions.Instance) int64 {
 	var (
 		realMandelboxCapacity int
 	)
@@ -26,16 +26,16 @@ func ComputeRealMandelboxCapacity(imageID string, activeInstances []subscription
 		}
 	}
 
-	return realMandelboxCapacity
+	return int64(realMandelboxCapacity)
 }
 
 // ComputeExpectedMandelboxCapacity is a helper function to compute the expected mandelbox capacity.
 // Expected capacity is the number of free mandelboxes available on instances with active status and
 // in starting instances.
-func ComputeExpectedMandelboxCapacity(imageID string, activeInstances []subscriptions.Instance, startingInstances []subscriptions.Instance) int {
+func ComputeExpectedMandelboxCapacity(imageID string, activeInstances []subscriptions.Instance, startingInstances []subscriptions.Instance) int64 {
 	var (
-		realMandelboxCapacity     int
-		expectedMandelboxCapacity int
+		realMandelboxCapacity     int64
+		expectedMandelboxCapacity int64
 	)
 
 	// Get the capacity from active instances
@@ -44,10 +44,10 @@ func ComputeExpectedMandelboxCapacity(imageID string, activeInstances []subscrip
 	// Loop over starting instances (status PRE_CONNECTION), only consider the ones with the current image
 	for _, instance := range startingInstances {
 		if instance.ImageID == imageID {
-			expectedMandelboxCapacity += int(instance.RemainingCapacity)
+			expectedMandelboxCapacity += instance.RemainingCapacity
 		}
 	}
 
 	expectedMandelboxCapacity += realMandelboxCapacity
-	return expectedMandelboxCapacity
+	return int64(expectedMandelboxCapacity)
 }
